requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
colorama==0.4.4
docopt==0.6.2
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pykwalify==1.8.0
pyparsing==3.0.7
pytest==7.0.1
python-dateutil==2.8.2
PyYAML==6.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
six==1.16.0
tomli==2.0.1
